Improve reliability of tests: handle missing pgrep and allow a grace period while checking for thread exit
This guarantees the test will not fail because of a lack of pgrep.  However, even with these changes, it could still fail waiting for threads to exit if the system is under heavy load.  The next checkin will fix that issue, but this checkin at least makes the test more reliable in most cases.