FIXED: Hang with multi-byte UTF8 characters reported by @Losbarthos
Details in: https://github.com/SWI-Prolog/packages-mqi/issues/14

The MQI implementation had a bug where sends to MQI expected the message header to provide a count of Unicode code points instead of what it should have been: UTF8 bytes. This has been corrected but is a breaking change, so:

ADDED: A simple protocol versioning interface to the MQI. There is now a mqi_version/2 predicate and the MQI returns Major and Minor version in its response to an initial connect message.

ADDED: The swiplserver Python library properly detects the original and fixed versions and works against both of them.

TEST: Added tests for multi-byte UTF8 characters